move out sentinel ut helpers
